refactor(management): decompose account inspection page
